Update markdown body contribution to use new contributions hook
