+++ workspaces/ui-v2/src/optic-components/documentation/MarkdownBodyContribution.tsx
@@ -1,10 +1,9 @@
 import * as React from 'react';
-import { ChangeEvent, useEffect, useState } from 'react';
+import { ChangeEvent } from 'react';
 import makeStyles from '@material-ui/styles/makeStyles';
 import { TextField } from '@material-ui/core';
 import classNames from 'classnames';
-import { useDebounce } from '../hooks/ui/useDebounceHook';
-import { useContributionEditing } from '../hooks/edit/Contributions';
+import { useContributionEditing, useValueWithStagedContributions } from '../hooks/edit/Contributions';
 import { OpticBlueReadable } from '../theme';
 import ReactMarkdown from 'react-markdown';
 
@@ -20,24 +19,11 @@ export function MarkdownBodyContribution({
   defaultText,
 }: MarkdownBodyContributionProps) {
   const {
-    lookupContribution,
     isEditing,
-    stagePendingContribution,
   } = useContributionEditing();
 
-  const value = lookupContribution(id, contributionKey);
   const classes = useStyles();
-
-  const [stagedValue, setStagedValue] = useState(value);
-
-  const debouncedChanges = useDebounce(stagedValue, 1000);
-
-  useEffect(() => {
-    if (debouncedChanges && stagedValue !== value) {
-      stagePendingContribution(id, contributionKey, debouncedChanges);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [debouncedChanges]);
+  const { value, setValue } = useValueWithStagedContributions(id, contributionKey);
 
   const inner = isEditing ? (
     <TextField
@@ -46,9 +32,9 @@ export function MarkdownBodyContribution({
       variant="filled"
       multiline
       placeholder={defaultText}
-      value={stagedValue}
+      value={value}
       onChange={(e: ChangeEvent<HTMLInputElement>) => {
-        setStagedValue(e.target.value);
+        setValue(e.target.value);
       }}
     />
   ) : (
